requirements.txt
streamlit
python-dotenv
langchain
langchain-core
langchain-openai
openai
tiktoken
